doubles to ints

+++ Lab6/BLOSUM.java
@@ -30,8 +30,8 @@ final public class BLOSUM {
     * Refer to above matrix for corresponding amino acids
     * i.e. score(A, R) corresponds to  matrix[0][1]=matrix[1][0]=-1
     */
-    public static double[][] getMatrix() {
-        double[][] matrix =  new double [][] {
+    public static int[][] getMatrix() {
+        int[][] matrix =  new int [][] {
             { 4, -1, -2, -2,  0, -1, -1,  0, -2, -1, -1, -1, -1, -2, -1,  1,  0, -3, -2,  0},
             {-1,  5,  0, -2, -3,  1,  0, -2,  0, -3, -2,  2, -1, -3, -2, -1, -1, -3, -2, -3},
             {-2,  0,  6,  1, -3,  0,  0,  0,  1, -3, -3,  0, -2, -3, -2,  1,  0, -4, -2, -3},
